OrganisationController - implemented read organisation endpoint

+++ src/Core/Givt.OnlineCheckout.API/Models/Organisations/GetOrganisationRequest.cs
@@ -2,6 +2,6 @@
 {
     public class GetOrganisationRequest
     {
-        public int OrganisationId { get; set; }
+        public long OrganisationId { get; set; }
     }
 }
